pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "risk-ads"
version = "0.1.0"
description = "Sistema de optimización de Google Ads para risktactical.mx"
requires-python = ">=3.12"
authors = [
    {name = "Efrain Ponce", email = "efrain.ponces@gmail.com"}
]

dependencies = [
    "google-ads>=24.1.0",
    "google-auth-oauthlib>=1.2.0",
    "duckdb>=1.1.0",
    "pandas>=2.2.0",
    "typer>=0.13.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-cov",
]

[project.scripts]
risk-ads = "src.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
